+++ rustq.exs
@@ -136,6 +136,7 @@ generate "native-stubs", "apps/gpui/lib/gpui/native_generated.ex" do
             MetaAST.function!(GPUI.Codegen.Native.TextBoundary, :text_buffer_undo),
           text_buffer_redo:
             MetaAST.function!(GPUI.Codegen.Native.TextBoundary, :text_buffer_redo),
+          host_info: MetaAST.function!(GPUI.Codegen.Native.RuntimeBoundary, :host_info),
           start_runtime:
             MetaAST.function!(GPUI.Codegen.Native.RuntimeBoundary, :start_runtime),
           stop_runtime: MetaAST.function!(GPUI.Codegen.Native.RuntimeBoundary, :stop_runtime),